Enter DOIs through a Shiny Module

Input, validate and return DOIs in a shiny app.

doiEntryApp()

doiEntryUI(id, width = "100%", ...)

doiEntryServer(id, char_limit = 100000L)

Arguments

id
The id string to be namespaced (optional).

width
The width of the input, e.g. '400px', or '100%'; see validateCssUnit().

...
Arguments passed on to shiny::textAreaInput

inputId

The input slot that will be used to access the value.

label

Display label for the control, or NULL for no label.

value

Initial value.

height

The height of the input, e.g. '400px', or '100%'; see validateCssUnit().

cols

Value of the visible character columns of the input, e.g. 80. This argument will only take effect if there is not a CSS width rule defined for this element; such a rule could come from the width argument of this function or from a containing page layout such as fluidPage().

rows

The value of the visible character rows of the input, e.g. 6. If the height argument is specified, height will take precedence in the browser's rendering.

placeholder

A character string giving the user a hint as to what can be entered into the control. Internet Explorer 8 and 9 do not support this option.

resize

Which directions the textarea box can be resized. Can be one of "both", "none", "vertical", and "horizontal". The default, NULL, will use the client browser's default setting for resizing textareas.

char_limit
Integer scalar, giving the maximum number of characters. To protect shiny against overlong strings, you can limit the maximum length of strings allowed. This limit is still enforced server-side, not client-side, so the protection is not bullet-proof.

Related Functions and Methods

Functions

doiEntryApp: Test app

doiEntryUI: Module UI

doiEntryServer: Module server

See also